requirements.txt
Flask==3.0.0
pdfplumber==0.11.0
PyPDF2==3.0.1
python-dotenv==1.0.0
Pillow>=11.0.0
openai>=2.0.0
gunicorn>=21.2.0
reportlab>=4.0.0
html2text>=2020.1.16

